+++ src/Stripe.net/Services/Cards/CardUpdateOptions.cs
@@ -6,10 +6,10 @@ namespace Stripe
     public class CardUpdateOptions : BaseOptions
     {
         [JsonProperty("exp_month")]
-        public int? ExpMonth { get; set; }
+        public long? ExpMonth { get; set; }
 
         [JsonProperty("exp_year")]
-        public int? ExpYear { get; set; }
+        public long? ExpYear { get; set; }
 
         [JsonProperty("name")]
         public string Name { get; set; }
