+++ src/Stripe.net/Services/PaymentIntents/PaymentIntentCaptureOptions.cs
@@ -6,10 +6,10 @@ namespace Stripe
     public class PaymentIntentCaptureOptions : BaseOptions
     {
         [JsonProperty("amount_to_capture")]
-        public int? AmountToCapture { get; set; }
+        public long? AmountToCapture { get; set; }
 
         [JsonProperty("application_fee_amount")]
-        public int? ApplicationFeeAmount { get; set; }
+        public long? ApplicationFeeAmount { get; set; }
 
         [JsonProperty("transfer_data")]
         public PaymentIntentTransferDataOptions TransferData { get; set; }
